setup-wizard: install: removed "-D" option
"-D" is not available for "install" tool on Mac OS X and thus breaks build.

As it is not necessary when target directories are created with $(INSTALL_DIR) beforehand, we don't need it here.

Anyway, the point of the $(INSTALL_XXX) macros is to abstract from using specific "install" options, so it seems cleaner to me that way.